backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
passlib[argon2]==1.7.4
sqlalchemy==2.0.45
asyncpg==0.31.0
pydantic-settings==2.12.0
email-validator==2.3.0
python-dotenv==1.2.1
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
aiosqlite==0.22.1
python-jose==3.5.0
python-multipart==0.0.22